Add files via upload
Project Description

This is a simple Python project for decrypting monoalphabetic ciphers. I used the tkinter library to make a small window where the user can enter the encrypted message and the 26-letter cipher alphabet.

When the user clicks the Decrypt button, the program will convert the encrypted text back into the original message using a substitution method. It also keeps the capital and small letters the same. If the user enters something wrong, like a cipher alphabet that’s not 26 letters, it shows an error.